requirements.txt
flask>=3.0.0
flask-socketio>=5.3.0
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
mediapipe>=0.10.0
scikit-learn>=1.3.0
scipy>=1.11.0
